requirements.txt
requests
pandas
python-dotenv

